+++ kiwi.js
@@ -1041,7 +1041,7 @@ exports.tools = require('./tools');
  * Version
  */
 
-exports.version = '0.2.0';
+exports.version = '0.2.1';
 
 
 /**